final debugging

+++ main.py
@@ -59,14 +59,74 @@ async def on_message(message):
                     if piece == 'o':
                         r.set((str(message.channel.id)+"garr"+"movec"), 'x')
                     # arr = io.BytesIO()
+                    cgarr = pickle.loads(r.get(str(message.channel.id)+"garr"))
+                    if(generate(pos, piece, cgarr) != -1):
+                        cgarr = generate(pos, piece, cgarr)
+                        r.set(str(message.channel.id) +
+                              "garr", pickle.dumps(cgarr))
+                        if piece == 'x':
+                            r.set((str(message.channel.id)+"garr"+"movec"), 'o')
+                        if piece == 'o':
+                            r.set((str(message.channel.id)+"garr"+"movec"), 'x')
+                        if check_win(cgarr, piece) == 1:
+                            del flushRedis(str(message.channel.id))
+                            if(piece == 'x'):
+                                await message.reply(discord.File("X final.gif"))
+                            if(piece == 'o'):
+                                await message.reply(discord.File("O final.gif"))
 
-                    img = render(garray)
-                    img.save(str(message.channel.id)+'.png')
-                    # arr.seek(0)
-                    await message.reply(file=discord.File(str(message.channel.id)+'.png'))
-                    os.remove(str(message.channel.id)+'.png')
+                        if check_draw(cgarr, piece) == 1:
+                            del flushRedis(str(message.channel.id))
+                            await message.reply("GAME IS A DRAW")
+                        if(check_win(cgarr, piece) == 0 and check_draw(cgarr, piece) == 0):
+                            img = render(garray)
+                            img.save(str(message.channel.id)+'.png')
+                            await message.reply(file=discord.File(str(message.channel.id)+'.png'))
+                            os.remove(str(message.channel.id)+'.png')
+                    if(generate(pos, piece, cgarr) == -1):
+                        await message.reply("occupied position,enter an empty position")
+
             '''if(r.get(str(message.channel.id)+"garr"+"movec").decode('utf=8') != piece):
                 await message.reply('game failure,botDev is checking')'''
+
+
+def check_win(cgarr, piece):
+    soln = [[1, 2, 3], [4, 5, 6], [7, 8, 9], [1, 4, 7],
+            [2, 5, 8], [3, 6, 9], [1, 5, 9], [3, 5, 7]]
+    if(piece == 'x'):
+        val = 0
+    if(piece == 'o'):
+        val = 1
+    c = 0
+    for i in cgarr:
+        for j in soln:
+            if(c == 3):
+                return 1
+            c = 0
+            for k in j:
+                if(cgarr[k-1] == val):
+                    c += 1
+                else:
+                    continue
+    return 0
+
+
+def check_draw(garr):
+    if(garr.count(0)+garr.count(1)):
+        return 1
+    else:
+        return 0
+
+
+def generate(pos, piece, garray):
+    if garray[pos-1] != -1:
+        return(-1)
+    else:
+        if(piece == 'x'):
+            garray[pos-1] == 0
+        if(piece == 'o'):
+            garray[pos-1] == 1
+        return(garray)
 
 
 def render(garray1):
@@ -123,8 +183,10 @@ def writeRedis(gArr):
     r.set(garray, pickle.dumps(gArr))
 
 
-def flushRedis():
-    r.flushdb
+def flushRedis(key):
+    r.delete(key)
+    r.delete(key+"garr")
+    r.delete(key+"garr"+"movec")
 
 
 keep_alive()
